separate interface in base and version specific
